requirements.txt
pytest==6.2.4
requests
flask
python-dotenv